The debugging in tension and compression works now.

+++ jmodels/jmodelYopiNew/jmodelyopi.h
@@ -81,7 +81,7 @@ namespace jmodels
         double un_ro;//reloading displacement
         double fm_ro; //reloading stress
         double un_hist_ten;
-        uint32 reloadFlag; //reloading flag
+        double reloadFlag; //reloading flag
         uint32 plasFlag;
         double dc_hist;
         uint32 pertFlag;
